1.025

+++ orders/get.php
@@ -55,6 +55,15 @@
 	}
 
    // 
+	if(isset($_GET['check'])) {
+		$id = strip_tags($_POST['id']);
+      $upd = db::query("UPDATE `retail_orders` SET `order_status` = 6 WHERE id = '$id'");
+      echo 'yes';
+
+      exit();
+	}
+
+   // 
 	if(isset($_GET['delete'])) {
       $id = strip_tags($_POST['id']);
       $del = db::query("DELETE FROM `retail_orders` where id = '$id'");
